Switch SheetChart/SheetImage dimensions from EMU to pixels
SheetChart.Width/Height, SheetImage.Width/Height, and
CellAnchor.ColumnOffset/RowOffset now take a double in pixels instead
of a long in EMU. The XLSX reader/writer converts to and from EMU at
the file boundary, and all internal pixel math in the overlays drops
its /9525 conversions. Matches the plain-pixel convention used by
ClosedXML, Telerik SpreadProcessing, and Syncfusion XlsIO; avoids
requiring users to think in a unit Excel itself hides.

+++ Radzen.Blazor/Spreadsheet/ResizeChartCommand.cs
@@ -8,11 +8,11 @@ public class ResizeChartCommand : ICommand
 {
     private readonly SheetChart chart;
     private readonly CellAnchor? newTo;
-    private readonly long newWidth;
-    private readonly long newHeight;
+    private readonly double newWidth;
+    private readonly double newHeight;
     private CellAnchor? oldTo;
-    private long oldWidth;
-    private long oldHeight;
+    private double oldWidth;
+    private double oldHeight;
 
     /// <summary>
     /// Creates a resize command for a TwoCellAnchor chart.
@@ -26,7 +26,7 @@ public class ResizeChartCommand : ICommand
     /// <summary>
     /// Creates a resize command for a OneCellAnchor chart.
     /// </summary>
-    public ResizeChartCommand(SheetChart chart, long newWidth, long newHeight)
+    public ResizeChartCommand(SheetChart chart, double newWidth, double newHeight)
     {
         this.chart = chart;
         this.newWidth = newWidth;
